fish: remove code host from checkout path, no longer needed
I had this setup because at one time I was cloning from a lot of disparate
sources, and this setup started as my $GOPATH.

However, this hasn't been useful as a $GOPATH for sometime (around the time
I dropped the src/ path component..), so might as well remove the code host
path component as well.

This also updates the `wk` command which previously depended on the code host being in the source path.